+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Gradle extension with configuration.
+  - Force mappings.
+  - Mappings provider (if you use another MC dev plugin).
 - Bon2Task for use BON2 in Gradle.
 
 [unreleased]: https://github.com/MJaroslav/Bon2Gradle/compare/9e79296...HEAD